@@ -50,7 +50,7 @@ internal class ImmutableList<T>
     {
         m_formatter = formatter;
         m_isISupportsReadonlyType = typeof(IImmutableObject).IsAssignableFrom(typeof(T));
-        m_list = new List<T>();
+        m_list = [];
     }
 
     /// <summary>
